+++ Funkcje_matematyczne/silnia.cpp
@@ -3,9 +3,7 @@
 
 using namespace std;
 
-int silnia(int n) {
-	//for (i = 1; i <= n; i++) {
-	// silnia * i;
+unsigned long long silnia(unsigned long long n) {
 	
 	if (n == 0) return 1;
 	else return n * silnia(n - 1);
